Añadida pila de fragments para poder usar el botón de atrás.
Más gestión de excepciones.
