+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/FeatherLyrics_APP.iml" filepath="$PROJECT_DIR$/FeatherLyrics_APP.iml" />
       <module fileurl="file://$PROJECT_DIR$/FeatherLyrics_APP_Poblenou.iml" filepath="$PROJECT_DIR$/FeatherLyrics_APP_Poblenou.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
