+++ app/src/main/java/tk/sbarjola/pa/featherlyricsapp/PerfilesUsuarios/BaseFragmentUser.java
@@ -6,6 +6,9 @@ import android.support.design.widget.TabLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.view.ViewPager;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
@@ -23,6 +26,13 @@ public class BaseFragmentUser extends Fragment {
     private TabLayout tabLayout;
     private ViewPager viewPager;
     private ViewPagerAdapter adapter;
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setHasOptionsMenu(true);
+    }
 
     @Nullable
     @Override
@@ -64,6 +74,32 @@ public class BaseFragmentUser extends Fragment {
             }
         });
     }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater){ //Afegim una opcio "Refresh" al menu del fragment
+        super.onCreateOptionsMenu(menu, inflater);
+
+        inflater.inflate(R.menu.local, menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.localMusic) {
+
+            ((MainActivity) getActivity()).abrirReproduccionesLocales();
+
+            return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
 
     public void addUser(String pagename) {
         Bundle bundle = new Bundle();
@@ -109,11 +145,10 @@ public class BaseFragmentUser extends Fragment {
 
     public void setupTabLayout() {
         selectedTabPosition = viewPager.getCurrentItem();
+
         for (int i = 0; i < tabLayout.getTabCount(); i++) {
             tabLayout.getTabAt(i).setCustomView(adapter.getTabView(i));
 
         }
-
-
     }
 }
