+++ app/src/main/res/menu/local.xml
@@ -2,8 +2,8 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/userProfile"
+        android:id="@+id/localMusic"
         app:showAsAction="ifRoom"
-        android:title="User Profile"
+        android:title="Local music"
         android:icon="@drawable/ic_person_pin_white_36dp"></item>
 </menu>
